// src/suppliers/dividendmax.com/types.ts
export interface SearchResult {
  flag: string; // us
  id: number; // 2793
  image: string; // https://cdn.dividendmax.com/companies/gladstone-investment-corporation.png
  name: string; // Gladstone Investment Corporation
  path: string; // /united-states/nasdaq/equity-investments/gladstone-investment-corporation/dividends
  search: string; // GAIN
  ticker: string; // GAIN
}